requirements.txt
cachetools==5.2.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
et-xmlfile==1.1.0
google-auth==2.11.0
google-auth-oauthlib==0.5.2
googletrans==3.1.0a0
gspread==5.5.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.23.2
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==1.26.12
